# frontend/openapi/admin-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REST to MCP Admin API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/tools/": {
      "get": {
        "tags": [
          "tools"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "isActive",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "mcpServerId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "mcpServerId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "slug": {
                            "type": "string",
                            "minLength": 1
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "inputSchema": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "outputSchema": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "examples": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                },
                                {
                                  "type": "boolean"
                                },
                                {
                                  "enum": [
                                    "null"
                                  ],
                                  "nullable": true
                                },
                                {
                                  "type": "array",
                                  "items": {}
                                },
                                {
                                  "type": "object",
                                  "additionalProperties": {}
                                }
                              ]
                            },
                            "nullable": true
                          },
                          "riskLevel": {
                            "type": "string"
                          },
                          "scopeIds": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "format": "uuid"
                            }
                          },
                          "mapping": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "backendApiId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "backendResourceId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "requestMapping": {
                                "type": "object",
                                "additionalProperties": {}
                              },
                              "responseMapping": {
                                "type": "object",
                                "additionalProperties": {}
                              },
                              "errorMapping": {
                                "type": "object",
                                "additionalProperties": {}
                              },
                              "authStrategy": {
                                "type": "string"
                              },
                              "timeoutOverrideMs": {
                                "type": "integer",
                                "exclusiveMinimum": true,
                                "minimum": 0,
                                "nullable": true
                              },
                              "retryOverride": {
                                "type": "object",
                                "additionalProperties": {},
                                "nullable": true
                              },
                              "isActive": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "backendApiId",
                              "backendResourceId",
                              "requestMapping",
                              "responseMapping",
                              "errorMapping",
                              "authStrategy",
                              "timeoutOverrideMs",
                              "retryOverride",
                              "isActive"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "mcpServerId",
                          "name",
                          "slug",
                          "title",
                          "inputSchema",
                          "outputSchema",
                          "examples",
                          "riskLevel",
                          "scopeIds",
                          "mapping",
                          "isActive"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0
                            },
                            "pageSize": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0
                            },
                            "total": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "pageCount": {
                              "type": "integer",
                              "minimum": 0
                            }
                          },
                          "required": [
                            "page",
                            "pageSize",
                            "total",
                            "pageCount"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "pagination"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mcpServerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "inputSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "outputSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "examples": {
                    "type": "array",
                    "items": {},
                    "default": []
                  },
                  "riskLevel": {
                    "type": "string",
                    "default": "low"
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  },
                  "scopeIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "default": []
                  },
                  "mapping": {
                    "type": "object",
                    "properties": {
                      "backendResourceId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "requestMapping": {
                        "type": "object",
                        "additionalProperties": {},
                        "default": {}
                      },
                      "responseMapping": {
                        "type": "object",
                        "additionalProperties": {},
                        "default": {}
                      },
                      "errorMapping": {
                        "type": "object",
                        "additionalProperties": {},
                        "default": {}
                      },
                      "authStrategy": {
                        "type": "string",
                        "default": "inherit"
                      },
                      "timeoutOverrideMs": {
                        "type": "integer",
                        "exclusiveMinimum": true,
                        "minimum": 0,
                        "nullable": true
                      },
                      "retryOverride": {
                        "type": "object",
                        "additionalProperties": {},
                        "nullable": true
                      },
                      "isActive": {
                        "type": "boolean",
                        "default": true
                      }
                    },
                    "required": [
                      "backendResourceId"
                    ],
                    "additionalProperties": false,
                    "nullable": true
                  }
                },
                "required": [
                  "mcpServerId",
                  "name",
                  "slug",
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "mcpServerId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "slug": {
                          "type": "string",
                          "minLength": 1
                        },
                        "title": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "inputSchema": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "outputSchema": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "examples": {
                          "type": "array",
                          "items": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "number"
                              },
                              {
                                "type": "boolean"
                              },
                              {
                                "enum": [
                                  "null"
                                ],
                                "nullable": true
                              },
                              {
                                "type": "array",
                                "items": {}
                              },
                              {
                                "type": "object",
                                "additionalProperties": {}
                              }
                            ]
                          },
                          "nullable": true
                        },
                        "riskLevel": {
                          "type": "string"
                        },
                        "scopeIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          }
                        },
                        "mapping": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "backendApiId": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "backendResourceId": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "requestMapping": {
                              "type": "object",
                              "additionalProperties": {}
                            },
                            "responseMapping": {
                              "type": "object",
                              "additionalProperties": {}
                            },
                            "errorMapping": {
                              "type": "object",
                              "additionalProperties": {}
                            },
                            "authStrategy": {
                              "type": "string"
                            },
                            "timeoutOverrideMs": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0,
                              "nullable": true
                            },
                            "retryOverride": {
                              "type": "object",
                              "additionalProperties": {},
                              "nullable": true
                            },
                            "isActive": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "backendApiId",
                            "backendResourceId",
                            "requestMapping",
                            "responseMapping",
                            "errorMapping",
                            "authStrategy",
                            "timeoutOverrideMs",
                            "retryOverride",
                            "isActive"
                          ],
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "mcpServerId",
                        "name",
                        "slug",
                        "title",
                        "inputSchema",
                        "outputSchema",
                        "examples",
                        "riskLevel",
                        "scopeIds",
                        "mapping",
                        "isActive"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/admin/v1"
    }
  ]
}